+++ react-backend/app.js
@@ -4,10 +4,14 @@ var path = require('path');
 var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 
+
+// Require Routes
 var indexRouter = require('./routes/index');
-var dataRouter = require('./routes/data');
+var dataRouter = require('./routes/dataRoute');
+
 
 var app = express();
+
 
 // view engine setup
 app.set('views', path.join(__dirname, 'views'));
@@ -19,6 +23,8 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
+
+// Use Routes
 app.use('/', indexRouter);
 app.use('/data', dataRouter);
 
@@ -37,5 +43,6 @@ app.use(function(err, req, res, next) {
   res.status(err.status || 500);
   res.render('error');
 });
+
 
 module.exports = app;
