+++ react-backend/models/db.js
@@ -1,0 +1,22 @@
+// Set up Default Database Connection
+
+var mysql      = require('mysql');
+var connection = mysql.createConnection({
+  host     : 'localhost',
+  user     : 'root',
+  port: 3306,
+  password : '',
+  database: "navaldb"
+});
+
+exports.connect = connection.connect(function(err) {
+  if (err) {
+    console.error('error connecting: ' + err.stack);
+    return;
+  }
+
+  console.log('connected as id ' + connection.threadId);
+});
+
+// Export DB connection
+module.exports = connection;
