--- react-backend/routes/data.js
@@ -1,39 +1,0 @@
-var express = require('express');
-var router = express.Router();
-
-
-var mysql      = require('mysql');
-var connection = mysql.createConnection({
-  host     : 'localhost',
-  user     : 'root',
-  port: 3306,
-  password : '',
-  database: "navaldb"
-});
-
-connection.connect(function(err) {
-  if (err) {
-    console.error('error connecting: ' + err.stack);
-    return;
-  }
-
-  console.log('connected as id ' + connection.threadId);
-});
-
-
-
-/* GET users listing. */
-router.get('/', function(req, res, next) {
-	// Comment out this line:
-  //res.send('respond with a resource');
-  connection.query('SELECT * from films', function (error, results, fields) {
-		if (error) throw error;
-    res.json(results);
-    s
-  });
-  // And insert something like this instead:
-
-});
-
-
-module.exports = router;
